Add EU support

+++ app.js
@@ -20,8 +20,8 @@ const encodeVariants = (variants) => {
     return encodeURIComponent(JSON.stringify(varObj).replace(/'/g, '"'))
 }
 
-const checkoutVariants = (profile, variants) => {
-    const checkoutData = {
+const getCheckoutData = (profile, region, variants) => {
+    const usCheckoutData = {
         'store_credit_id': '',
         'from_mobile': '1',
         'cookie-sub': encodeVariants(variants),
@@ -42,6 +42,35 @@ const checkoutVariants = (profile, variants) => {
         'order[terms]': '1'
     }
 
+    const euCheckoutData = {
+        'store_credit_id': '',
+        'from_mobile': '1',
+        'cookie-sub': encodeVariants(variants),
+        'same_as_billing_address': '1',
+        'order[billing_name]': profile['billing_name'],
+        'order[email]': profile['email'],
+        'order[tel]': profile['tel'],
+        'order[billing_address]': profile['billing_address'],
+        'order[billing_address_2]': profile['billing_address_2'],
+        'order[billing_address_3]': profile['billing_address_3'],
+        'order[billing_zip]': profile['billing_zip'],
+        'order[billing_city]': profile['billing_city'],
+        'order[billing_country]': profile['billing_country'],
+        'credit_card[type]': profile['card_type'],
+        'credit_card[cnb]': profile['card_cnb'],
+        'credit_card[month]': profile['card_month'],
+        'credit_card[year]': profile['card_year'],
+        'credit_card[vval]': profile['card_cvv'],
+        'order[terms]': '1'
+    }
+
+    return region === 'us' ? usCheckoutData : euCheckoutData
+}
+
+
+const checkoutVariants = (profile, region, variants) => {
+    const checkoutData = getCheckoutData(profile, region, variants)
+
     return client.post(checkoutEndpoint, qs.stringify(checkoutData)).then(res => {
         return res.data
     }).catch(err => {
@@ -49,7 +78,7 @@ const checkoutVariants = (profile, variants) => {
     })
 }
 
-const exampleProfile = {
+const usExampleProfile = {
     'billing_name': 'Riley Brown',
     'email': 'someguy@gmail.com',
     'tel': '347-123-2000',
@@ -59,6 +88,24 @@ const exampleProfile = {
     'billing_city': 'New York City',
     'billing_state': 'NY',
     'billing_country': 'USA',
+    'card_cnb': '4424 9068 7255 8398',
+    'card_month': '01',
+    'card_year': '2020',
+    'card_cvv': '232',
+    'variants': ["1000", "20000"]
+}
+
+const euExampleProfile = {
+    'billing_name': 'Riley Brown',
+    'email': 'someguy@gmail.com',
+    'tel': '347-123-2000',
+    'billing_address': 'Watermanstraat 66',
+    'billing_address_2': '',
+    'billing_address_3': '',
+    'billing_zip': '5015 TG',
+    'billing_city': 'Tilburg',
+    'billing_country': 'NL',
+    'card_type': 'american_express',
     'card_cnb': '4424 9068 7255 8398',
     'card_month': '01',
     'card_year': '2020',
@@ -81,14 +128,14 @@ const checkStatus = (slug) => {
     })
 }
 
-const checkoutItems = (profile, variants) => {
+const checkoutItems = (profile, region, variants) => {
     logger.info(`Attempting to checkout with variants: ${variants}`)
-    checkoutVariants(profile, variants).then(res => {
+    checkoutVariants(profile, region, variants).then(res => {
         if(res.status === 'failed') {
             logger.fatal('Failed to checkout')
         } else if(res.status === 'outOfStock') {
             logger.warn('Item is out of stock.. Trying again in 1s')
-            setTimeout(() => checkoutItems(profile, variants), 1000)
+            setTimeout(() => checkoutItems(profile, region, variants), 1000)
         } else if(res.status === 'queued') {
             logger.info('In queue!')
             checkStatus(res.slug)
@@ -98,10 +145,32 @@ const checkoutItems = (profile, variants) => {
 
 
 (() => {
-    const profile = JSON.parse(fs.readFileSync('config.json'))
-    logger.info('Parsed config.json! Deleting it from disk.')
-    fs.unlinkSync('config.json')
-    if(profile) {
-        checkoutItems(profile, profile['variants'])
+    const args = process.argv;
+    let region = "us";
+
+    if(args.length > 2) {
+        if(args[2].toLowerCase() === 'us' || args[2].toLowerCase() === 'eu') {
+            region = args[2].toLowerCase();
+            logger.info(`Region set to ${region}`)
+        }
+    } else if(args.length <= 2) {
+        logger.info('Region defaulting to US. ')
     }
+
+    if (fs.existsSync('config.json')) {
+        const profile = JSON.parse(fs.readFileSync('config.json'))
+        logger.info('Parsed config.json! Deleting it from disk.')
+        fs.unlinkSync('config.json')
+        if(profile) {
+            checkoutItems(profile, region, profile['variants'])
+        }
+    } else {
+        fs.writeFile('config.json', JSON.stringify(region === 'us' ? usExampleProfile : euExampleProfile,null,2), { flag: 'wx' }, (err) => {
+            if (err) 
+                return console.log(err)
+
+            console.log(`Created config with example data in config.json`)
+        });
+    }
+    
 })()
